backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.7.0
pymongo==4.10.1
openai==1.58.1
qdrant-client==1.12.1
pdfplumber==0.11.4
python-docx==1.1.2
langchain-text-splitters==0.3.4
python-multipart==0.0.20
pydantic-settings==2.7.0
